libvirt: Update to 7.0.0
Upstream libvirt switched to meson in version 6.7.0 so the update
requires a bit more changes.

The main difference is that with meson we cannot override specific
binaries like we could with autoconf using ac_cv_path_* so we will
patch meson.build file to look for correct binaries.

Some of the options were not converted to meson:

    --with-driver-modules - it was not possible to disable even with
                            autoconf

Some of the options were dropped between 6.6.0 and 6.7.0:

    --without-hal - the HAL support was completely dropped (6.8.0)

    --without-dbus - libdbus was replaced by glib2 dbus (6.8.0)

    --without-virtualport - the code is compiled only on linux (6.9.0)

    --without-macvtap - the code is compiled only on linux if libnl
                        is available (6.9.0)

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>