py-dnspython: change back py27 subport's distfile
The zip version of the distfile was not mirrored after the subport was
changed to use it in 5747864, because no build was scheduled, because
the version did not change. This causes fetch problems for older
systems, and there's no reason to mirror the distfile in two formats
anyway.

Also make better use of portgroup defaults.

Fixes: https://trac.macports.org/ticket/62317