portgroups: use configure.build_arch
$build_arch contains the arch that the user configured in macports.conf
or the default arch for the current platform if none was configured,
while ${configure.build_arch} contains the arch that the port is
being built for. The two can differ if the port's supported_archs do
not include $build_arch. In that case, the arch can undergo a
"demotion" to a different arch supported by the port that will run on
the current platform, e.g. arm64 to x86_64, or x86_64 to i386.