grace: Fix implicit declaration of functions
Closes: https://trac.macports.org/ticket/61261

Fix netcdf variant so that it actually uses netcdf by removing a patch
of unknown purpose that has been in the port since the netcdf variant
was added:

See: https://trac.macports.org/ticket/16910

Fix port to ensure that netcdf support is not included unless the netcdf
variant is selected.

Add size to checksums.