py-pyobjc: don't use MacPorts libffi on Big Sur
PyObjC requires changes to libffi which Apple has yet to merge into
upstream. The build system incorrectly picked up the headers for the
system libffi, yet linked against MacPorts libffi, generating a binary
that failed with an error on load.

This fixes that by forcing use of the system-provided libffi on Big
Sur, and adjusts patches the sources to use the correct headers
elswehere. As libffi is a dependency of CPython itself, merely
removing it from the list of dependencies doesn't suffice, and we have
to include it by path.

While at it, I've applied some patches to force setuptools so that we
can build in parallel.

Fixes: https://trac.macports.org/ticket/62475