blueutil: Update to 2.8.0
Closes: https://trac.macports.org/ticket/62514

Use the makefile portgroup to use the right compiler and -arch flags.

Don't patch the makefile to change the prefix since the makefile is set
up to accommodate changing the prefix via command line argument. Do
patch the makefile to remove -Werror which we never want and to append
to MacPorts flags rather than overwrite them.

Since the build fails for i386 and since the port already requires 10.9
or greater, remove i386, ppc, and ppc64 from supported_archs.