icu: don't use runConfigureICU MacOSX
runConfigureICU is a shell script that
tries to set up some options for the build
that we don't need, as we already have these
set properly in base.

on occasion, this script results in multiple conflicting
flags being set

offers no benefits for MacPorts on MacOSX/macOS.

Not sure about freebsd -- impossible for me to test --
so leave it as is there.

no revbump as icu update coming shortly