icu: fix universal build on Darwin20+
prior to 10.14, universal means different library bit depths or endianness,
so we had to build separately and lipo together using the muniversal PG

We can't easily use the muniversal PG on darwin20+ as during the build,
icu has to run binaries to generate data, and the Intel Mac cannot run the
arm64 binaries generated by the muniversal PortGroup.

However we don't need to use the muniversal PortGroup  because the universal
build on Darwin20+ has the same library bit depth and endianness on both
Intel and arm64.

Therefore we can build universal in one pass using clang's ability to generate both archs at once.

closes: https://trac.macports.org/ticket/62548

no revbump as
1. this should change no files on disk already installed
2. a new version of icu is coming shortly