tiff: Don't use autoreconf
It was added years ago when tiff bundled a version of libtool that did
not pass the -stdlib flag on to the compiler. tiff now bundles a newer
version of libtool that doesn't have that problem.

Redo patchfiles to patch configure instead of configure.ac and
acinclude.m4 and rename patchfiles to indicate their purpose.