ttcp: Fix implicit declaration of functions
Fix implicit declaration of functions that prevented building with Xcode
12 and later.

Closes: https://trac.macports.org/ticket/62861

Add a Makefile and use it to build and install. Use the makefile 1.0
portgroup to use the right compiler, cflags, ldflags, and arch flags,
and to add a universal variant. Since this changes how it builds,
increase the revision.

Modernize checksums.

Indicate the license.

Switch the homepage to something that exists.