llvm/clang/lldb-12: new port
although this looks similar to the previous llvm ports,
in fact it is completely redone from scratch using
the cmake-1.1 portgroup, redoing all the patches in a
dedicated patch repo, and retooling the working parts.

The polly variant for llvm has been rolled into the standard
build, as it appears to build on all systems, and there seems
to be no reason why you would not want it.

the flang port does not automatically require gcc (although it
will require gcc to work). It seemed too heavy a burden to put
on users to install the entire gcc setup for clang. Soon, flang
will not require gcc at all to function.

I considered rolling the clang variants into the standard build,
but in the end, I did not, and left them as variants.

Most things (llvm, clang, compiler-rt, etc) have tests that work now.

I would like to add the BOOTSTRAP feature to the build on a
future update, to build a maximally functional clang compiler,
but this version, sans-BOOTSTRAP, is the way we have build
things all these years.

Systems < 10.14 have been temporarily restricted from building clang
and the related projects until these have been fully verified
and then the build on these older systems will be re-enabled.

The rewrite and move to cmake-1.1 might introduce some new issues that
I have not as yet foreseen. Please open a ticket if you spot
such an issue and I'll try to take care of it promptly.

As much cruft as I could release has been eliminated from this
build, and although I was hoping for an even cleaner Portfile, this
is what I thought we could accept for the time being.