php73: Avoid need for newer compiler on El Capitan
Most versions of Xcode clang on El Capitan miscompile the datetime
parsers that were generated with re2c 2.0.3 resulting in a runtime hang.
Rather than blacklisting Xcode clang for this reason, regenerate the
parsers with re2c 0.15.3. Installing and using re2c 0.15.3 is faster
than installing a new version of clang.

See: https://trac.macports.org/ticket/61351

Certain details about the compiler that was used to compile php are
recorded in php_config.h. This causes problems if the compiler used to
compile php is not the same as the compiler used to compile a php
extension later.

Closes: https://trac.macports.org/ticket/62022