clang-3.4: remove lion workaround
at one time clang-3.4 required ICU on these older systems,
and ICU would not build with the system compiler on lion,
so a workaround was added. That workaround was later removed when
the need for ICU was removed (for older systems).

So this block can be restored to it's previous state.