clang-11,12,devel: default analyzer on darwin 11+
the analyzer variant adds a dependency on perl5 at runtime

Although at present MacPorts compiler selection builds perl5
using clang-9.0, which is fine and builds a fully-capable perl5,
it is possible that at some point in the future MacPorts might change to
default to a newer clang like clang-10, 11 or newer on 10.6 and earlier.

If that should happen, then there would be circular dependancy.

So we will remove the analyzer variant from all clang versions on systems
where the system roots can't build a fully capable perl5, in case
at some point in the future the MacPorts defaults might change.

closes: https://trac.macports.org/ticket/63265