gcc7-bootstrap: new port
this port is to be part of a solution to bootstrapping
older systems. It is as self-contained as I can make it,
relying on only system roots to build on all systems,
and using no other ports in MacPorts to build or run.

The aim is to make it immune to changes in Portfiles over time.

This port will be used to build gcc11 directly. libgcc11
requires a c++11 compiler to build.

This port will also be used to build a version of clang,
perhaps clang-3.8 or clang-4.0, directly against libstdc++,
bypassing libc++. This "bootstrap clang" can then be used
to build a fully-capable libc++ to be installed, and also can
be used then to build any "clang dependency" ports that are
needed, without worrying about the path to a capable clang.