dibbler: 1.0.2RC1-20200723
This is not a stable release, but the developer states that the project
is concluded, so no further releases are expected. Instead of the last
stable release 1.0.1 or the last release candidate 1.0.2RC1, we may as
well take the latest commit from the repo since it contains further
fixes including a fix for newer macOS versions.

Closes: https://trac.macports.org/ticket/52018
Closes: https://trac.macports.org/ticket/55143