unbound: ability to run multiple unbounds in parallel
The startup item will start/stop an unbound daemon for each
unbound*.conf file in the /opt/local/etc/unbound directory.
If there is only a single unbound*.conf file it behaves as before.
This is useful for being able to run rspamd properly. Without
this, the rspamd-initiated DNS qeuries might be blocked. Using
a second non-forwarding unbound solves the issue.