gdbm: more selective patch for Tiger
although off_t is defined in the sys/types.h header, on
Leopard and newer, off_t is selectively defined in stdio.h  without
including the entire sys/types.h header.

It likely makes little overall difference, but including the entire
sys/types.h header could have unexpected effects, so we will duplicate
the behaviour of Leopard and newer  as much as possible.

This patch is always applied as is desired, but blocked on systems
where the deployment target is newer than 10.4.

This version still builds of course, passes the test suite, and
successfully builds ports that require gdbm such as perl5.34.

As this change is installed in the gdbm.h header, revbump needed.