ntpsec: update to 1.2.1
This includes the patches for compatibility with macOS<10.13,
which can also be seen (more readably) at:

    https://gitlab.com/fhgwright/ntpsec/tree/macports_1_2_1

Up until recently, 10.5 x86_64 wasn't possible due to broken
dependencies.  That case now builds and passes its tests, but doesn't
actually work.  Pending resolution of this issue, x86_64 is excluded
from supported_archs for 10.5, as well as for 10.4 since it might have
the same issue and hasn't been tested.  At present (on 10.5 x86_64),
this results in at least one broken universal dependency, so it's back
to not being buildable at all.  This is no worse than previous
versions.

TESTED:
Tested (including building the usual set of variant combinations)
on 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15 x86_64.