ruby26: Fix build issues on 10.4, 10.5.
This includes a patch for some OS version issues common to earlier
Ruby versions:

The OS version conditional in dln.c is comparing an undefined macro to
a sometimes undefined (and essentially useless) macro.  However, the
relevant code doesn't seem to be built for any OS version, making the
bug irrelevant.  Nevertheless, it's now fixed, since another, similar
fix was needed, anyway, and it might become relevant if something else
changes.

The OS version conditionals in error.c were using another useless
macro for the comparison, this time with a real effect, though only in
a couple of messages related to crash reports on < 10.6.  This is now
fixed.

It also includes an updated patch for a 10.4 issue:

The context register naming issue on 10.4 was sort of fixed upstream
but tied to an inappropriate config flag.  The patched version uses an
SDK version macro to make the decision.

A newly discovered issue, common to some older Ruby versions, is:

The 10.5 x86_64 build fails if libunwind-headers is active.  Since
this is an unusual configuration, and the need for libunwind-headers
seems to be waning, it's not worth fixing "right".  This just adds the
usual failure for such cases.  It's unknown why this only affects 10.5
x86_64 (and perhaps 10.4 x86_64, though that's untested).

The remaining issues were new in ruby26:

The syntax-checking compiler invocations in transform_mjit_header.rb
require a C99 or later compiler, but don't include compiler flags from
the build.  There's code to detect the non-C99 case and add -std=c99
to the real uses, but it relies on matching the error message from the
test case.  Recent versions of gcc use locale-dependent quotes which
may break such comparisons, depending on the current locale settings.
The fix recommended by gcc documentation is to run it with LC_ALL=C to
get traditional ASCII quotes.  This adds code to apply that
environment to the compiler test invocation.

For some strange reason, using gcc-4.2 for Ruby's coroutine-related
assembler code on x86_64 passes "-arch i386" to the assembler, thereby
causing failures on x86_64 register references.  To fix that, cases
with gcc-4.2 and x86_64 targeted now pass the correct architecture in
ASFLAGS.

Also corrects a small typo in the description.

Also fixes some lint errors in the Portfile.

Since this affects the installed code for < 10.6, it includes a
revbump.

Universal builds still fail on 10.4 i386 due to a weird configure bug,
and 10.4 ppc has broken universal dependencies.

TESTED:
Built -universal on 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15
x86_64.  Built +universal on 10.5-10.13.