xhyve: add missed dependencies
`xhyve` is linked to `zlib`. Usually it is installed as side effect from
another dependencies but it isn't trur for 10.11 and 10.10 builds. This
builds are failed on linking phase
```
Undefined symbols for architecture x86_64:
  "_deflate", referenced from:
      _rfb_send_screen in xhyve_lto.o
      _rfb_send_all in xhyve_lto.o
  "_crc32", referenced from:
      _rfb_send_screen in xhyve_lto.o
  "_deflateEnd", referenced from:
      _rfb_thr in xhyve_lto.o
      _rfb_send_screen in xhyve_lto.o
      _rfb_send_all in xhyve_lto.o
  "_deflateInit_", referenced from:
      _rfb_thr in xhyve_lto.o
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
```

Closes: https://trac.macports.org/ticket/57292