asciidoctor: Revert to last fully working version of Ruby (2.6)
Ruby3.0 doesn't build for 10.4, 10.5, 10.7, or 10.8.  Ruby2.7 also
fails for 10.4, 10.5, and 10.7.  The Ruby dependency shouldn't be changed
until the new Ruby version is fully compatible.

Also adds myself as a maintainer, mainly to keep these sorts of changes
from slipping through in the future.

TESTED:
Built and ran on macOS 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15
x86_64.