openssl: Fix build on 10.4, RNG on 10.10-10.11
The patch for pre-Yosemite OS versions relied on Availability.h, which
didn't exist in 10.4.  This updates the patch by substituting the
__ENVIRONMENT... versions of the macros to avoid the need for the
header, as well as eliminating the superfluous defined() checks.

In addition, the RNG from Apple crypto apparently doesn't work
correctly prior to 10.12, so this updates the version threshold for
its use.  Because this changes code on 10.10 and 10.11, it includes a
revbump.  Dependents should only require revbumps to pick up the fix
if they link statically.

It also renames and adjusts the patch to conform to MacPorts style,
as well as adjusting the comments related to the patch.

Closes: https://trac.macports.org/ticket/63421

TESTED:
Built on 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15 x86_64.
Verified that libcrypto references _CCRandomGenerateBytes in OS
versions >=10.12 and not otherwise.
Tests pass on 10.4-10.5 and 10.9-10.15.
Test framework has preexisting issues on 10.6-10.8.