openssl11: Fix commit IDs in Portfile comments
The upstream commit IDs referenced in relation to the pre-Sierra patch
were for the master branch.  This replaces them with the versions for
the OpenSSL_1_1_1-stable branch, matching the correct versions of the
code.

This potentially affects the openssl port, but currently ignores it
due to the other pending change.  The upstream code may be different
in the 3.0.0 case, anyway.

TESTED:
Passes port lint.  Comment-only change.