gpsd: update to 3.23.1
TESTED:
Tested (including building the usual set of variant combinations and
running the tests) on 10.5 ppc, 10.5-10.6 i386, and 10.6-10.15 x86_64.
Some variant combinations were excluded in some cases due to issues
with the dependencies, though it's expected that they'd work if the
broken dependencies were fixed.
Not tested on 11.x.
Known not to build on 10.4.