yubico-authenticator: Add missing dependencies
The yubico-authenticator binary links against qt5-qtmultimedia, but does
not declare this dependency:

| $> for file in $(otool -L work/destroot/Applications/MacPorts/Yubico\ Authenticator.app/Contents/MacOS/yubioath-desktop | sed 1d | awk '{print $1}' | grep '/opt/local'); do port -q provides "$file"; done | sort -u
| qt5-qtbase
| qt5-qtdeclarative
| qt5-qtmultimedia
| qt5-qtquickcontrols2

Additionally, yubico-authenticator fails to start without
qt5-qtquickcontrols installed:

| QQmlApplicationEngine failed to load component
| qrc:/qml/main.qml:454:5: Type Navigator unavailable
| qrc:/qml/Navigator.qml:350:9: Type NewCredentialView unavailable
| qrc:/qml/NewCredentialView.qml:6:1: module "QtQuick.Dialogs" is not installed
| Segmentation fault: 11

Add these dependencies to fix the issues. Bump the revision so that the
dependencies that are recorded in the registry are updated for everyone.
Without this, people could uninstall qt5-qtmultimedia or
qt5-qtquickcontrols and would end up with a broken yubico-authenticator.