ntpsec: Fix openssl shim conflict.
This adds an ugly wscript patch to ensure that the OpenSSL version
selection is honored in the build, thereby avoiding a sensitivity to
the shim port's selection.

Also enables "port test", though the normal build procedure runs the
tests, anyway.

Renames the patchfile to something more general.

The revbump may not be strictly necessary, but is included to track
the updated patch.

TESTED:
Ran build/test/install with the usual variant combinations, on
10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15 x86_64.