py-cycler: Restore compatibility with Python <3.6
The upstream code dropped support for Python <3.6.  So far, undoing
this only involves a few small changes, so it's patched here.  If this
becomes unmanageable at a later time, it may become necessary to pin
the version for older Python versions.

TESTED:
Tested with Python 2.7 and 3.5-3.10, on 10.4-10.5 ppc, 10.4-10.6 i386,
and 10.5-10.15 x86_64.  Passed in all cases that didn't involve broken
dependencies, including all 10.6-10.15 cases.