openssl*: Restore versioned libs at top level
The libraries with versioned names can be installed in the top-level
${prefix}/lib directory without conflict.  Doing this (at least for
openssl11) allows the existing dependent installs to continue working.
The unversioned libraries still need to be provided by the shim port,
pointing to only the one version chosen as the default by the shim
port.

Due to the transfer of ownership of a top-level library, a transient
activation failure may occur.  There doesn't seem to be a good fix for
this, though a forced activation or out-of order activation fixes it.

TESTED:
Built, installed, and tested on 10.4-10.5 ppc, 10.4-10.6 i386, and
10.5-10.15 x86_64.  The openssl11 tests fail on 10.6-10.8 as before.
The openssl3 tests also fail on 10.6-10.8, as well as on 10.10-10.11,
which is almost certainly unrelated to this change.
The preexisting ntpsec @1.2.1_0 install now works again.