p5-pod-parser: new port
This is a legacy distribution which was included in perl core
through perl5.30 but has been removed in perl5.32 and later.
Provided here to allow older perl code that assumes it's in core
to build on these newer versions.  New code is urged to use
its modern replacement Pod::Simple. See p5-pdl as a case in point.