openvdb: rewrite Portfile, add new variants, enable features
This commit is a fairly significant rewrite of the Portfile. After digging more deeply into the OpenVDB source code and documentation, it turns out that the Portfile is only performing a very minimal build. There are some features, such as support for hardware intrinsics, which are strongly recommended by the upstream authors, but are not enabled by default. I have added variants so that users can control some of these options from the 'port' command.

I have added the new jemalloc variant to the default variants, since using jemalloc is strongly recommended by the upstream devs. I have added the new docs variant to the default variants, as suggested by the Best Practices section of the MacPorts Guide.

In addition, I'm not sure why building the Python bindings is the default variant. Currently, the two ports that depend on openvdb are openimageio and blender. OpenImageIO doesn't need the Python module. I will need to check to see whether Blender somehow uses pyopenvdb (I'm doubtful: I believe that Blender links directly to libopenvdb.dylib at build time, which would likely be more performant than making calls to the Python module at run time), but I currently believe that the Python module is unused. I may remove python from the default variants list in the future, if it turns out that nothing is making use of it.