ksh93: Fix build due to memccpy misdetection
Closes: https://trac.macports.org/ticket/43392

Also suppress implicit function declaration errors that cause the build to
fail with Xcode 12 and later.

Add most remaining patches from Apple's ksh 93u+ build on the theory that
they must be good for something.

Since we're forcing a rebuild anyway, take the opportunity to refine the
version number so that "93u+" is included since some users may be looking
for that.

No longer deprecated now that development is continuing in the 93u+m fork.