php-xhprof: Update to "revived" fork with version 2.3.7
Based on "official" fork that is used by PECL itself.

This has been tested with php72-php81 (self tests and some real apps)
and it's working ok so far. Note it only includes the minimum changes
required, not sure if there is anything else to modify (ports versions
or things like that).

TODO: Maybe the opposite conflict should be added to the tideways
package as part of this too?