mod_tile: Handle conflict with boost port
This port depends on the `mapnik` port, which builds with Boost.

This application determines how `mapnik` was built during `configure`
via an M4 macro that makes calls to `mapnik-config` (installed with
the `mapnik` port) and adds appropriate flags for the libs and include
parameters, which will specify the correct Boost version for the
build.

The build failed if the `boost` port was installed.  This Portfile now
specicies the same Boost version as the `mapnik` port, ensuring the
correct version of Boost is used.