python27: fix support for virtualenv on macOS 12
On macOS 12 (Monterey) NSLibraryNameForModule returns the real path of
the dynamic library instead of the path to the symlink in a virtualenv.
This breaks support for virtualenv.

By applying the same patch as Apple does for its Python version, which
no longer uses the deprecated NSLibraryNameForModule function, support
for virtualenv is restored.

See also:
https://github.com/apple-oss-distributions/python/blob/python-136.120.2/2.7/fix/getpath.c.ed