python pg: add python.pep517_backend option
This automatically adds appropriate build dependencies for common
PEP 517 back-ends. The default is setuptools, matching the fallback
behaviour recommended by the PEP and implemented in py-build. The
option can be cleared or set to something unrecognised like "other" if
you don't want back-end deps added. This is useful if you are using
a back-end that the portgroup doesn't know about, or for ports that
install a back-end that can build itself.

Thanks to reneeotten for the idea.