freac: Don't invoke xcodebuild
The build system only uses xcodebuild to determine the Xcode version
on the basis of which it decides what flags to use during a release
build but we are not doing a release build.

xcodebuild requires Xcode and MacPorts hides Xcode from the build system
because the port does not specify "use_xcode yes" so the build log is
otherwise filled with errors about xcodebuild not being able to run and
subsequent script errors from the build system which did not expect
that to occur.

Closes: https://trac.macports.org/ticket/64281