ngspice: update version to 35 and new ngspice-lib subport
Unfortunately, shared library and the binary are mutually exclusive.
Since I haven't found a viable way to patch this behavior I create a
subport to build and install the dynamic library.
The other way, is to build two times the same source with different
flags but this is generally not macports compliant.