llvm-devel: Update portfile to mirror developments made to production clang portfiles.
A number of improvements have been made to the production clang port files which
have so far not been migrated to the llvm-devel suite of ports. This updates the
port to be be based off the current llvm-13 portfile, with then the minimal changes
needed to build a development snapshot. Should make it much easier in the future
to cut future production llvm-X ports from this development port (assuming it is
kept up to date).

Also update to latest 20220103 commit.

Add myself as a maintainer.

Fixes: https://trac.macports.org/ticket/63715