wolfssl: Use the requested number of build jobs
wolfssl's build system defaults to using a number of jobs that is 1 more
than the number of CPUs. Tell the build system not to do that ("disable
jobserver") so that the number of jobs the user has requested in
macports.conf is used instead.

MacPorts doesn't use a -j flag to run the test phase in parallel but
this build system normally does, so continue to do that since this build
system has been designed to accommodate it.