squish: fix dylib version number
The version-numbered dylib that gets generated is currently hardcoded to "0.0", which results in a versioned dylib of 'libsquish.0.0.dylib'. This commit fixes the incorrect usage of the dylib version number; it will now start generating the versioned dylib as 'libsquish.1.15.dylib'. I have left the "SOVERSION 0.0" in place in the 'CMakeLists.txt' file, so that 'libsquish.0.0.dylib' will still get generated as a (symlink) compatibility version; this means that revbumps to dependent ports are not needed.