warzone2100: add license_noconflict
The warzone2100 subport is getting flagged with a '"warzone2100" is not distributable because its license "gpl" conflicts with license "OpenSSL" of dependency "openssl11"' on the buildbot builders. However, warzone2100 is in fact NOT dependent on openssl11; the curl port has already been transitioned to using openssl3, and the port that is still dependent on openssl11 is asciidoctor, one of the tools that warzone2100 is using during the build process.