gnu-apl: Add deps and fix macOS 11+ libtool bug
Fix license.

Add deps on fftw-3, pcre2, sqlite3, and xorg-libxcb which are used
opportunistically.

Ensure that gtk3 is not used. It is also found opportunistically but
causes the build to fail because the code uses a Linux-specific
constant. --without-gtk3 did not work so I prevent finding pkg-config.

Fix the well-known libtool bug that caused build failure on macOS 11+.
This configure script was old enough that it also had the Yosemite
libtool bug which caused builds on macOS 10.10-10.15 to erroneously use
the flat namespace.