py-wxpython-4.0: use internal wxWidgets version
since wxPython is using a commit and not a tagged version of wxWidgets
it is better to use and install that version. This avoid strange issues
or crashes particularly with complex software like Kicad that maintains
sessions between wxWidgets library and the python realm.

To use that wxWidgets version you need to use the following path for
wx-config:

  set python_version  3.9
  set py_ver_no_dot   [join [split ${python_version} "."] ""]
  set python_framework_dir ${frameworks_dir}/Python.framework/Versions/${python_version}

  ${python_framework_dir}/bin/wx-config

In cmake

  -DwxWidgets_CONFIG_EXECUTABLE=${python_framework_dir}/bin/wx-config