py-requests-file: Fix checksum mismatch
Revert part of the preceding commit and go back to using the github
portgroup for now to fix the checksum mismatch that the preceding commit
caused (by causing the distfile to fetch from pypi which serves a
different file than GitHub). The github portgroup can be removed (and
fetching can go back to using pypi) if desired when the port is updated
to a new version.

Closes: https://trac.macports.org/ticket/64538