godot: use external libs
This commit adds support for compiling Godot using external dependencies. It is nearly identical to the patch that has been implemented by one of the maintainers of the Godot package for Homebrew, as discussed here:

https://github.com/godotengine/godot-proposals/issues/1796

I will also be submitting an upstream PR to get this feature incorporated into the Godot code base.