php-pcov: new port (from PECL)
This extension is orders or magniture quicker than the coverage
facilities provided by XDebug (both v2 and v3) and provide
similar results. When testing large code-bases it really makes
a difference. So I think it may be useful to have it available.

I've been using it already some time with PHP 73/74/80 and it's
working perfectly, so far.

I'm not sure about which is the policy about incompatibilities
and how this should be applied, but this is worth reading:

https://github.com/krakjoe/pcov#interoperability

And that's all, ciao :-)