oof2: update to 2.2.2
This *should* fix the last of the build problems caused by
buffer overflows in our ancient build tools.  I finally figured
out how to test it in situ, using the full path names used by
macports, and not the shorter ones that it uses when building
from a local repository.