golang-1.0: Do not fail when pkgs conflict on prefix
A port that has both gopkg.in/src-d/go-git.v4 and
gopkg.in/src-d/go-git-fixtures.v3 in its go.vendors would previously
fail, because no sha1_short was given for such ports, and src-d-go-git
is a prefix of src-d-go-git-fixtures. This caused the glob expression in
the post-extract to match both folders, move both folders, and then fail
when processing the second package.

Resolve this by only choosing the match from the glob expression that
does not contain dashes in the part that matched the wildcard.