gdb: add patch to prevent lock-ups when running a binary
Workaround for this bug https://sourceware.org/bugzilla/show_bug.cgi?id=24069
which causes GDB to hang when running a binary. gdb may still hang, but
a ^C (Control-C) will unlock it.
While the root cause is still unknown and the patch included here is only a
workaround, the patch has reportedly a near-100% success rate in preventing
lock-ups so it's still an improvement on the status quo.

Let's keep an eye on the bug and revert this commit once one of 2 things
happen:
 - this patch gets merged into master
 - the root cause is debugged and the proper fix is shipped from
   upstream