arrangodb: fix broken Portfile
The `target_architecture` variable is undeclared on any platform other than
i386.

Additionally, the Portfile asserts that it only supports x86_64 (via
`supported_archs`), so it's not likely that this port would ever be built on
i386, which is the only scenario where `target_architecture` would be set.

Fixed by eliminating the variable and directly appending the config option for
the appropriate platform.