gcc10-bootstrap: use flat patches
Idea to use muniversal PG and arch-specific patches wasn't good that
leads to broken build due to not applied patches on 10.8..10.13,
and may produce unpredicted result when ppc and i386 is mixed.

Here, patches a bit reorganized to be applied all together,
that allows to use the same sources to build it.

It also includes backported version a fix for
https://trac.macports.org/ticket/63161 and all macOS 12-related patches
are put into one.