php-xmlrpc: new port (from PECL)
Before PHP 8.0, the xmlrpc extension was bundled with PHP, but
now it's moved to PECL.

https://wiki.php.net/rfc/unbundle_xmlprc
https://php.watch/versions/8.0/xmlrpc

Current version (1.0.0RC3) works ok for PHP 8.0 and 8.1. While it's
really using old libraries and the recommendation is to move away
from it to newer/better alternatives, there is still some code out
there using it.

Hence, this provides the extension, built from PECL, for PHP 8.0 and up.