rrdtool: Use release tarball; fixes version number
The previous 1.8.0 version of this port used an automatically generated
tarball. The git repository contains an outdated configure script and
associated files that were not updated for the 1.8.0 release so the
program misidentified itself as 1.7.2, among who knows how many other
unintended differences. By switching to the release tarball we pick up
the version number change and any other bugfixes or changes that the
developers made in configure.ac, m4 files, and Makefile.in files.

https://github.com/oetiker/rrdtool-1.x/issues/1163

Also fix implicit declaration of functions in configure script.

https://github.com/oetiker/rrdtool-1.x/pull/1162