cltversion: fix spurious warnings with no CLT installation on macOS ≥ 11
On macOS 11 and 12, when Command Line Tools (CLT) are not installed,
cltversion produced these warnings:

Warning: cltversion: The Command Line Tools are installed, but MacPorts cannot determine the version.
Warning: cltversion: For a possible fix, please see: https://trac.macports.org/wiki/ProblemHotlist#reinstall-clt

This was confusing, because CLT were not installed. The linked document
discusses reinstalling CLT, but this is nonsense when CLT is not present
and there’s no need or desire to install CLT. CLT may intentionally be
absent if a fully-functioning Xcode installation is available. This
warning may have caused quite a bit of busywork by macOS 11 and 12
users, dutifully silencing the warning by installing CLT unnecessarily.

The warning appeared because cltversion looked for
/usr/lib/libxcselect.dylib on disk to determine if it should look for a
new-style CLT installation, but /usr/lib/libxcselect.dylib does not
exist as a file on disk on macOS 11 or later (Darwin 20 or later), since
individual system dylibs are only shipped in the dyld shared cache on
these newer OS versions. The file’s absence caused cltversion to search
for an old-style CLT installation, checking for the existence of
/usr/bin/make. On modern OS versions, /usr/bin/make is always present as
a stub that delegates to make in the active developer tools
installation—it’s bundled with the OS and will never be missing.

This fix makes cltversion look for a new-style CLT installation if the
OS is so new that it doesn’t ship system dylibs outside of the dyld
cache. Rest assured that on OS versions this new, no old-style CLT
installations are even possible.